test: Split open-ended supersession regression into its own file
observation_service_test.go exceeded the 1500-line service-convention
limit after adding the DAT-2 open-ended supersession regression test.
Move that test to observation_supersession_test.go in the same package;
no behavioural change.
